requirements.txt
streamlit
sentence-transformers
faiss-cpu
groq
pypdf
python-dotenv
gtts
speechrecognition
pyaudio
